Enable force mode compatibility with various move types (#230)
This should make force mode work with various of move types. 
With this commit, the joint positions used for checks and interpolations
is now done without force mode modifications.

When the robot is in an offset situation, where the robot isn't at the position that
is currently used for motion commands, this will use the non-offset position to
check whether a new motion is within the reachable limits of the robot.