CLEANME Use TCP_NODELAY for incoming connections on ReverseInterface
We are sending packets of size 32 Bytes to the reverse interface. When no
TCP_NODELAY option is enabled, the OS might group packets and send multiple
at once at a later time.

If our sending frequency and our timeout are setup in a way that delaying
packets by 2 or three cycles hits that timeout, this might happen due to
packet grouping.

Using TCP_NODELAY should prevent that from happening.